requirements.txt
asgiref==3.7.2
Django==5.0
django-humanize==0.1.2
humanize==4.9.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
